refactor(#792): split volume_invariants_tests.rs by responsibility (2282->251 fixtures + 5 concern files, PINNED verbatim)
Sections A-P grouped into concern files: chain invariants, split-mono/preset,
spectral audit, elastic-ring audit, broadcast/format. Tests moved verbatim (pinned).
Shared fixtures (chain builders, bare_chain_for, pink_noise) in base. Off debt.
Full engine suite: 611 pass, 0 fail.
